refactor: move DTO classes to dto package for better organization
- LottoTickets
- LottoResult
