+++ src/main/java/lotto/dto/LottoResult.java
@@ -1,7 +1,8 @@
-package lotto.model;
+package lotto.dto;
 
 import java.math.BigDecimal;
 import java.util.Map;
+import lotto.model.LottoRank;
 
 public record LottoResult(
         Map<LottoRank, Integer> ranks,
