+++ src/main/java/lotto/dto/LottoTickets.java
@@ -1,6 +1,7 @@
-package lotto.model;
+package lotto.dto;
 
 import java.util.List;
+import lotto.model.Lotto;
 
 public record LottoTickets(
         List<Lotto> tickets
