+++ src/test/java/lotto/model/LottoResultCalculatorTest.java
@@ -11,6 +11,8 @@ import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.List;
 import java.util.Map;
+import lotto.dto.LottoResult;
+import lotto.dto.LottoTickets;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
